% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukceh_colours.R
\docType{data}
\name{ukceh_colours}
\alias{ukceh_colours}
\title{UKCEH palette colours}
\format{
A linked list, names of useful colours palettes
and their associates colours as a character string:
}
\source{
\url{https://brandroom.ceh.ac.uk/sites/default/files/2022-01/UKCEH_Colour_Guidelines.pdf}
}
\usage{
ukceh_colours
}
\description{
A list of useful colours paletters for UKCEH.
}
\keyword{datasets}
