% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukceh_get_colours.R
\name{ukceh_get_colours}
\alias{ukceh_get_colours}
\title{Request a UKCEH colour palette}
\usage{
ukceh_get_colours(palette_name = NA)
}
\arguments{
\item{palette_name}{the name of the colour palette}
}
\value{
string of hex colour codes
}
\description{
Request a UKCEH colour palette
}
\examples{
# Selecting the logo colours
ukceh_get_colours("logo")

# Colours are stored in `ukceh_colours`
# You can call names(ukceh_colours) to see all options
}
