% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukceh_pick_theme.R
\name{ukceh_pick_theme}
\alias{ukceh_pick_theme}
\title{Return a UKCEH theme}
\usage{
ukceh_pick_theme(theme = NULL)
}
\arguments{
\item{theme}{name of a UKCEH theme.}
}
\value{
a specified UKCEH bslib theme
}
\description{
The result of this function should be used as the \code{theme} argument for
\code{\link[shiny]{bootstrapPage}}, \code{\link[shiny]{fluidPage}},
\code{\link[shiny]{navbarPage}}, or \code{\link[shiny]{fixedPage}}.
}
\examples{
if(interactive()) {
library(ukcehthemes)
shinyApp(
  ui = fluidPage(theme = ukceh_pick_theme("generic"),
    ...
  ),
  server = function(input, output) { }
)
}
}
