% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukceh_titlePanel.R
\name{ukceh_titlePanel}
\alias{ukceh_titlePanel}
\title{UKCEH titlePanel replacement}
\usage{
ukceh_titlePanel(title = "UKCEH Shiny app", windowTitle = title)
}
\arguments{
\item{title}{}

\item{windowTitle}{}
}
\value{

}
\description{
UKCEH titlePanel replacement
}
\examples{
\dontrun{
ui <- fluidPage(
theme = UKCEH_theme,
UKCEH_titlePanel("My first Shiny app"),
sidebarLayout(
  sidebarPanel(
    ...
  ),
  mainPanel(
    ...
  )
)
)

}
}
