% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcehtheme.R
\name{ukcehtheme}
\alias{ukcehtheme}
\title{Return a UKCEH theme}
\usage{
ukcehtheme(theme = NULL)
}
\arguments{
\item{theme}{Name of a UKCEH theme.}
}
\value{

}
\description{
The result of this function should be used as the \code{theme} argument for
\code{\link[shiny]{bootstrapPage}}, \code{\link[shiny]{fluidPage}},
\code{\link[shiny]{navbarPage}}, or \code{\link[shiny]{fixedPage}}.
}
\examples{
\dontrun{
shinyApp(
  ui = fluidPage(theme = ukcehtheme("generic"),
    ...
  ),
  server = function(input, output) { }
)
}
}
