% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukceh_titlePanel.R
\name{ukceh_titlePanel}
\alias{ukceh_titlePanel}
\title{UKCEH titlePanel replacement}
\usage{
ukceh_titlePanel(apptitle = "UKCEH Shiny app")
}
\arguments{
\item{apptitle}{the title of the app itself.}
}
\value{

}
\description{
UKCEH titlePanel replacement
}
\details{
this function replaces titlePanel() and adds a title in the format
"\code{apptitle} | UK Centre for Ecology and Hydrology", the UKCEH logo and a
favicon for the browser tab.
}
\examples{
\dontrun{
ui <- fluidPage(
theme = UKCEH_theme,
UKCEH_titlePanel("My first Shiny app"),
sidebarLayout(
  sidebarPanel(
    ...
  ),
  mainPanel(
    ...
  )
)
)

}
}
